feat: finished with storage,add storage ,information ,test

+++ database/migrations/2021_03_06_232540_create_storages_table.php
@@ -20,7 +20,7 @@ class CreateStoragesTable extends Migration
             $table->text("description")->nullable();
             $table->integer("sku");
             $table->string("mm");
-            $table->integer("price");
+            $table->float("price");
             $table->integer("quantity")->default(0);
             $table->timestamps();
         });
